+++ Models/Produto.cs
@@ -5,11 +5,11 @@ namespace sgp.Models
   [Serializable]
   public class Produto
   {
-    public long Codigo { get; set; }
+    public int Codigo { get; set; }
     public string Nome { get; set; }
     public double Preco { get; set; }
 
-    public Produto(long codigo, string nome, double preco)
+    public Produto(int codigo, string nome, double preco)
     {
       this.Codigo = codigo;
       this.Nome = nome;
